upgrade to .net 8 (#53)
* upgrade to .net 8

* resolve comments

* resolve comments

* Upgrade dotnet-sonarscanner version to 6.0.0

This commit updates the version of the dotnet-sonarscanner package from 5.13.1 to 6.0.0 in the _build.csproj file. The upgrade is done to take advantage of the new features and fixes included in the newer version.

* Enable Unsafe Binary Formatter Serialization

Added the <EnableUnsafeBinaryFormatterSerialization> tag set to true in _build.csproj file. This was necessary to allow the serialization of binary format data which is not safe by default.

---------

Co-authored-by: Gian-Luca Polenta <gpolenta@techgroup.ch>
Co-authored-by: Oleg Shevchenko <olsh.me@gmail.com>